packages/core/src/analyzer-legacy/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Core dependencies from the original app
python-arango>=7.5.0
gitpython>=3.1.32
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.0.0

# HTTP client for webhooks and integrations
httpx>=0.25.0

# Additional utilities
psutil>=5.9.0

# Optional dependencies
plotly>=5.15.0
jira>=3.4.0
matplotlib>=3.7.0
seaborn>=0.12.0